like-service/requirements.txt
uvicorn===0.27.0.post1
fastapi===0.109.0
psycopg2-binary===2.9.9
python-consul===1.1.0
sqlalchemy===2.0.30

